backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy>=2.0.25
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1  # OAuth 2.0 client for social media APIs

# HTTP Clients and API Communication
httpx>=0.24.0,<0.25.0
requests==2.31.0  # Fallback HTTP client, used by some OAuth libraries

# Retry Logic and Resilience (Critical for API reliability)
tenacity==8.2.3  # Retry decorator with exponential backoff for API calls

# Rate Limiting (Important for API compliance)
limits==3.6.0  # Rate limiting library for API calls
cachetools==5.3.2  # Used by limits for caching

# Date/Time Utilities
python-dateutil==2.8.2  # Better date parsing for API responses

# HTML/XML Parsing (For web scraping and content extraction)
beautifulsoup4==4.12.2  # HTML parsing for website content
lxml==4.9.3  # Fast XML/HTML parser (alternative to html.parser)
html5lib==1.1  # HTML5 parser for BeautifulSoup

# Background Jobs and Task Queue
redis==5.0.1
rq==1.15.1
apscheduler==3.10.4

# Environment and Configuration
python-dotenv==1.0.0

# External Services
supabase==2.3.0

# Email and Messaging Utilities
email-validator==2.1.0  # Email validation for prospect data

# URL and Domain Utilities
tldextract==5.1.0  # Extract TLD, domain, and subdomain from URLs
urllib3==2.1.0  # HTTP library with connection pooling

# JSON and Data Processing
orjson==3.9.10  # Fast JSON library (optional but recommended for performance)

# Logging and Monitoring
structlog==23.2.0  # Structured logging (optional but recommended)

# Social Media API Specific (if using official SDKs)
# Note: We're using direct API calls, but these can be useful:
# linkedin-api==2.0.0  # LinkedIn SDK (optional)
# instagrapi==2.0.0  # Instagram API wrapper (optional, use with caution)

# Web Scraping and Content Extraction
selenium==4.15.2  # Browser automation (optional, for complex scraping)
webdriver-manager==4.0.1  # Selenium driver management

# Data Validation and Sanitization
bleach==6.1.0  # HTML sanitization for user-generated content
validators==0.22.0  # URL and email validation

# Async Utilities
aiofiles==23.2.1  # Async file operations
asyncio-throttle==1.0.2  # Rate limiting for async operations

# Error Handling and Resilience
backoff==2.2.1  # Alternative retry library (complement to tenacity)
